Delinearizator inherits from CXref rather than XRef. This is due to encryption credentials manipulation methods which are not available in XRef (therefore we have to inherit publicly and not protected). This also means that we can't deallocate stream in destructor because this is already done in CXref destructor.
delinearize method checks for credentials and if they are necessary and not
present, we will simply return with an error with logged error message.

Delinearization::delinearize fails for encrypted documents because we should
not (for legal and ethic) reasons provide unencrypted content to the
delinearized files. We can enable it when encrypted conetent writing is
implemented.