From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH @num@/@total@] writeObject cleanup
OldstylePdfWrite currently uses two functions for writing objects:
	- writeIndirectObject
	- writeDirectObject
Their implementations are almost the same, thus a lot of code duplication.

This patch removes them and provides unified writeObject function which covers
both of them. Implementation is based on the writeIndirectObject and updated to
add indirect object specific stuff only if required by parameter.

streamToCharBuffer signature has to be updated too because we want to
prevent fake ref creation for direct object streams (btw. why is this
function in the cdict.h?).


* cdict.h, cobject2xpdf.cc
	- streamToCharBuffer signature changed Ref -> Ref* (ignored if
	  asIndirect==false)
* pdfwriter.cc
	- writeIndirectObject, writeDirectObject removed
	- writeObject added