From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH @num@/@total@] FilterStreamWriter introduction
This patch adds new FilterStreamWriter class which is responsible for proper
writing of stream objects. Currently we are writing streams in non
compressed form without any filters. This is far from ideal situation
because even simple change in the content stream may enlarge the final size
very rapidly.

FilterStreamWriter class is the first step to change this behavior. In the
first step we have simply created a class which provides interface for all
available filters (FilterStreamWriter) and current stream writing logic was
moved to the first implementation of filter stream writer named
NullFilterStreamWriter.

To provide easy-to-use and transparent interface we are using the following
hierarchy and class relations:
- FilterStreamWriter
	- provides factory method getInstance which returns proper filter
	  writer instance based on the given stream object
	- stream writers can be registered in runtime by
	  registerFilterStreamWriter method
	- if getInstance is not able to find any suitable writer for a given
	  object, it will try default writer
	- if even default doesn't support given object, it will fall back to
	  the NullFilterStreamWriter which writes raw data without any filters.

- default filter stream writer is useful for give-at-least some filter if
  the required one is not available (e.g. LZWDecode can be replaced by
  FlateDecode or stack of filters can be replaced by only one of them).

- each filter stream implementation is required to provide:
	- supportObject - it understands how to write filtered data
	  specified by object.
	- compress - write data from given object to the stream. Note that
	  this method can use different filters than specified in the
	  object, but it definitely has to write valid object (e.g. stack of
	  filters can be reduced or one filter can be replaced by another).


* FilterStreamWriter, NullFilterStreamWriter classes added
* writeObject
	- updated to use FilterStreamWriter rather than direct code for
	  stream writing


Changelog v1 -> v2
==================
* design documentation updated