Current implementation of bufferFromStreamData seems to be absolutely broken when we get rid of the old stream filters logic. This also fixes the bug 270.
At first, we can't use the stream length (coming from stream dictionary) for
upper read characters count bound because this value marks the length of
_encoded_ (number of bytes after stream keyword and the last stream byte)
stream. So that we have to read until stream eof and realloc buffer if it
is not sufficient.

Secondly, we can't use BaseStream for data retrieving because this will
return underlaying data source without any filters. However this function is
expected to provide decoded data from stream which are later used by filters
to be encoded. This bug is rather nasty, because you can't observe it during
simple testing (modify document and save) because every changed stream
contains no filters and so that base stream and stream are essentially the
same. However delinearization which also uses this function will refer to
unchanged object (BaseStream is FileStream and Stream might or might not be
some kind of FilterStream).

Original implementation worked because we didn't processed returned data (no
filter has been implemented). If we had some filters then we would encode
(somehow encoded data) with filters type information got lost.