There are 2 reasons for this chnage:
* Current content stream modification code produces 0 length streams
  (sometimes and I am not sure exactly when). Moreover it forces updating
  all streams from the content stream which is changed.
* Some pdfs can really contain 0 length streams when they are decoded. I
  don't see clear reason but it is reality (e.g. PDFReference16.pdf and
  object 4129). When we are delinearizing such a document, then we decode
  this stream and as it seems to be zero length

This results in situation when we want to write 0 length streams which is
considered to be an error in object zlib stream object writing path[*].

Unfortunately this leads to a corrupted document in the end, because objects
are not written, but xref contains entries for them.

Note that this is _NOT_ a fix for the first issue, because it has to be
fixed in content stream manipulation layer, but it provides workaround to
prevent from documents corrupting. We should provide correct fix ASAP.


[*] It doesn't makes sense to write streams with 0 length - they don't hold
any useful information. If we want to get rid of an stream, we can simply
remove it from the contentstream's array.

* kernel/pdfwriter.cc
	- ZlibFilterStreamWriter::deflate - skip compression when the
	  bufferFromStreamData returns an empty buffer (0 length stream)