From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH @num@/@total@] kernel/filters remove
kernel/filters infrastructure is not used anywhere in the kernel/gui tree,
so it can be removed. It seems that iostreams don't helped much in stream
encoding and we can live without them quite well.

This patch removes kernel/filters.{cc,h} as well as dependency on
boost_iostreams package.

* configure.in
	- Boost iostreams detection removed
* README updated
* doc/user/install/requiredlib.xml
	- updated
*
* kernel/Makefile/filters.{cc,h}
	- removed
* kernel/Makefile
	- HEADERS, SOURCES updated to not include filters.{cc,h}
* tests/kernel/testcstream.cc
	- include to filters.h removed
	- code depends on Printable functor so it was moved from filters
	  (Is this code correct?)