From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH @num@/@total@] NullFilterStreamWriter with unchanged data
Reimplementation of NullFilterStreamWriter to write original stream data
rather than decoded data. This makes sense especially for cases when we want
to write unchanged stream object (e.g. Delinearizator code path) so that we
don't have to compress and can just copy data as they are.

This is very easy now, as we have bufferFromStream method (where we simply
use baseStream) and we need just to do some house keeping around this
function.

* NullFilterStreamWriter::null_extractor added
	- simple wrapper around bufferFromStream which will use base stream
	  so it reads data without any decompression in the path.
* NullFilterStreamWriter::compress
	- uses null_extractor rather than convertStreamToDecodedData