From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH @num@/@total@] ZlibFilterStreamWriter::support reimplementation
ZlibFilterStreamWriter should not be so invasive and it should be used only
if there is no filter for stream or the filter is already FlateDecode.
This change is important for paths when we write unchanged stream objects
(e.g. in Delinearization) where ZlibFilter would overwrite original filters.

This is not a problem, but it is not optimal, as we have to compress all
data even though they are already compressed with different filter.

* ZlibFilterStreamWriter::support
	- checks stream dictionary and makes supported only those streams
	  which have no filters or FlateDecode