From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH @num@/@total@] Filters registration cleanup
Filter writers interface is defined but not used anywhere. Filter writers
initialization is hardcoded in the pdfwriter module.

This will finish this approach and integrate filter writers implementators
into the pdfedit_core_dev_init function. To enable this, we have to export
NullFilterStreamWriter and ZlibFilterStreamWriter interface to the
pdfwriter.h header file.

Diff for pdfwriter.cc looks rather messy, because diff wasn't able to cope
with this kind of change, but no method has been changed by this patch.

Besides this changes, FilterStreamWriter interface is updated to contain
unregisterFilterStreamWriter and clearAllFilterStreamWriters methods to
enable correct cleanup.

* pdfwriter.h
	- ZlibFilterStreamWriter, NullFilterStreamWriter exported here
* pdfwriter.cc
	- FilterStreamWriter::unregisterFilterStreamWriter,
	  FilterStreamWriter::clearAllFilterStreamWriters added
	- only implementation from ZlibFilterStreamWriter and
	  NullFilterStreamWriter kept here
* pdfedit-core-dev.cc
	- pdfedit_core_dev_init initializes filter writers
	- pdfedit_core_dev_destroy cleanup for filter writers