From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH @num@/@total@] Factory for pdfoperators
Currently we have only SimpleGenericOperator and UnknownCompositeOperators
operators classes and we are using simple instantiation. This is, however,
far from the best solution, because we have problem when we introduce new
specialized operators (like the one in the following patch).

Therefore createOperator factory function was introduced. I wasn't sure
whether this function can be placed as static in the PdfOperator class, so
it is standalone at the moment. I can move it there if you want Jozo.

I have run kernel_tests TEST_PDFOPERS TEST_CCONTENTSTREAM for zadani.pdf and
pdf specification without any problems. So this change should be safe.


* gui/base.cc
	- Base::createOperator uses factory function for operators rather
	  than hardcoded SimpleGenericOperator

* kernel/ccontentstream.cc
	- createOperator renamed createOperatorFromStream (to distinguish it
	  from factory method) and updated to use pdfoperator factory function
	  rather than hard-coded instance creation (with exception of inline
	  images which need special treatment). Most of the original code moved
	  to the pdfoperator.cc
	- createOperands renamed to createOperandsFromStream (to be in sync
	  with createOperatorFromStream)
	- checkAndFix renamed to checkAndFixOperator and moved to podfoperator
	  module

* kernel/pdfoperators.{cc,h}
	- createOperator added
	- checkAndFixOperator added