From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH @num@/@total@] TextSimpleOperator introduction
This patch introduces the TextSimpleOperator specialized class for text
operators holding a text and replaces script based getTextFromTextOperator
function implementation by native TextSimpleOperator::getRawText function.

We need this new operator for proper handling of encoding for text objects.
QSPdfOperator::getEncodedText is the definitive interface for scripting to
provide correctly encoded text used by getTextFromTextOperator scripting
function.
Currently we are simply providing raw text without any consideration of font
encoding. This patch prepares background for proper implementation inside
the TextSimpleOperator class.

I also think that it is cleaner to have getTextFromTextOperator implemented
in the core rather than scripting because this functionality may be required
by some code which is not based on scripting (e.g. users of devel package)
and it is too core to be scriptable (but this can be arguable)


* gui/pdfoperators.qs
	- getTextFromTextOperator implementation removed and replaced by
	  simple call to getEncodedText

* gui/qspdfoperator.{cc,h}
	- QSPdfOperator::getEncodedText added to enable
	  TextSimpleOperator::getEncodedText method to be exported to the
	  scripting.
	  I am not sure whether this is proper place, but there doesn't seem
	  to better one.

* kernel/pdfoperators.{cc,h}
	- TextSimpleOperator class added
		- inherits from SimpleGenericOperator
		- adds getRawText
			- based on previous getTextFromTextOperator
			  implementation and doesn't implement the encoding
			  conversion yet
	- createOperator factory function updated to create TextSimpleOperator
	  for text operators

* kernel/stateupdater.h
	- isTextOp added

Changelog v2 -> v1
==================
* getEncodedText ranamed to getRawText because we don't implement any
  encoding related stuff in this patch, so this way it should be more clear
  Note that QSPdfOperator::getEncodedText name is kept, because it really
  involves encoding set for application.
* createOperator factory updated