From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH @num@/@total@] Encoded text string setting
TextSimpleOperator can't be self-contained - in terms of proper
initialization because we don't have enough information during instance
creation (in constructor) so that we have to wait for StateUpdater which
knows current GfxState for each pdf operator and misuse it for
TextSimpleOperator after-initialization.

[
I don't like that very much, because this is error prone (we can forget
to call initEncodedText for some text operators - this is basically copied
code spread around op*Update functions). But this seems like the only way
how to implement it.
]

All data required for later GfxFont instance retrieving are stored in the
FontData structure now (initialized by setFontData from state updater). This
way getFontData can get font instance anytime and use it for rawText
decoding (as well as opposite direction encoded -> raw text transformation).

* kernel/pdfoperator.{cc,h}
	- FontData added - defined in cc file because we don't need to have
	  this interface public
	- TextSimpleOperator
		- fontData added - keeps all necessary data for font
		  identification
		- getCurrentFont added - returns valid GfxFont instance for
		  associated font from resources associated with content
		  stream
		- setFontData, getFontData added

* kernel/stateupdater.cc
	- opTjUpdate, opSlashUpdate, opTJUpdate
		call setFontData

* gui/qspdfoperator.cc
	- getEncodedText uses getFontText rather than getRawText

Changelog v1 -> v2
==================
* getFontText, getFontName, setFontData added
* get rid of cached rawString and encodedString
	- they may be useful and can be added later, but there is no real
	  point in having them now
* initEncodedString removed (calls replaced by setFontData) and font encoding
  transformation code moved to getFontText
* don't initialize encoded text directly, but provide GfxFont and initialize
  all necessary data to get GfxFont later when needed
	- initEncodedtext removed and replaced by setFontData