From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH @num@/@total@] fake_install_tools introduction
This patch adds tools/fake_install_tools.sh script which aimed to replace
all tools which are used during {un}installation process. Its only purpose
is to fake operation given as the first argument and log this operation with
all parameters to the file.

This can be very handy when we want to check install targets and want to
make sure that nothing is broken (I was hit several times during my
experiments).

How it works?
To enable these fake tools, we just need to give --with-fake-install=YES
parameter to the configure and that's it. It will inject all variables
holding tools for installation with $(ROOT)/tools/fake_install_tools.sh
script and nothing more has to be done.

* configure.in
	- --with-fake-install parameter added (NO by default)
* tools/fake_install_tools.sh
	- script added