From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH] Check also for array filters in ZlibFilterStreamWriter::supportObject
Current implementation of ZlibFilterStreamWriter::supportObject checks only
for no Filters and a single filter. It doesn't check for the array of
filters.
This can be problem because:
	- array filters with a single filter inside is quite often situation
	- we store filters array with single FlateDecode into streams when
	  we are saving them through ZlibFilterStreamWriter

The first one is not such a ass pain because unchanged objects don't need
any special treatment and can be saved by NullFilterStreamWriter and changed
streams don't contain any filters anyway.

Nevertheless checking also array case is more logical here and we should do
it.


* getFiltersFromStream added
	- alternative for CStream::getFilters method working on the
	  low-level xpdf Object (this could be stored somewhere else maybe)
* ZlibFilterStreamWriter::supportObject
	- uses getFiltersFromStream rather than hardcoded isNull || isName
	  cases