From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH] Add fake page into mergeDialog
This patch adds one fake page at the end of the document page list to enable
simpler adding after the last page (because we are always inserting before
the currently selected page). Previously we had to select last page in the
document page list, then we had to insert page and move it down behind the
last page.

With the fake page we can simply select [END] page and insert a page before.
Down bottun is not active for any page immediately before the fake page so
it is always the last one in the list.

This patch implements issue http://pdfedit.petricek.net/bt/view.php?id=285