From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH] Fix lazy content stream parsing in CPageContents
This patch fixes bt#261

CPageContents keeps all content streams in _ccs variable which is not
initiailized in the constructor but rather in lazy fashion later when it is
needed in CPageContents::init method (parsing is rather complex operation and
we don't need content streams right after CPage resp. CPageContents is
created).

This, however, is not implemented correctly because not all public methods
call init method before they access _ccs so there are possible problems with
access to uninitialized content streams error (like the above issue).
Therefore we have to call init in each function - except for parse and
reparse - which uses _ccs. This should be correct because init doesn't parse
already parsed streams.