Fix for bt#306
Original report:

When building this in the openSUSE Build Service as a RPM package for openSUSE
11.1, post-build-checks complains about the following:

I: Program causes undefined operation
 (likely same variable used twiceand post/pre incremented in the same
expression).
 e.g. x = x++; Split it in two operations.
E: pdfedit sequence-point ../engine/qsnumber_object.cpp:85

I have attached a patch for that issue. Also it is fixed in newer versions of
QSA.


I think that this particular construct is not harmfull, but it is worth fixing
becuase it is a bad programming style.