From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH] Memory leaks in CXref::setCredentials
CXref::setCredentials may leak ownerPassword and userPassword on some return
paths from checkEncryptionCred (reported by Jozo).

The main problem is that checkEncryptionCred doesn't deallocate
ownerPassword and userPassword on some paths. Both strings are deallocated
only if SecurityHandler::checkEncryption method is called by
checkEncryptionCred:
-> SecurityHandler::checkEncryption
  -> makeAuthData
  -> freeAuthData

This means that they are not deallocated either when document is not
encrypted or security handler couldn't be found.

The other problem is that checkEncryptionCred doesn't report correctly all
possible cases. Currently we are not able to distinguish failed
authorization and none matching security handler! This is not a memory leak
problem but we end up with unexpected behavior.


* checkEncryptionCred
	- deallocate ownerPassword, userPassword if document is not
	  encrypted and if no matching security handler could be found
	- doesn't deallocate security handler if authorization fails to
	  distinguish the situation when no security handler could be found
* CXref::setCredentials
	- correctly handle no security handler case