From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH] FoFiTrueType::cvtSfnts potential mem leak fix
As reported by Jozo, vmtxTab can end up not deallocated. The report was
generated by automated checker tool and I have no more information about
condition. Nevertheless the spaghetti code in this function can lead to
really surprising results, so it is safer to deallocate vmtxTab at the just
to be sure (it is initialized to NULL so this should be safe).