From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH] CPdf destructor split up
Currently the whole cleanup work is done in CPdf::~CPdf which is not a bad
thing to do but it leads to a subtle side effect that most objects which are
cached (pages, indirect objects, ...) and they are wrapped by smart pointer
are released later than CPdf itself or at least not sooner than CPdf::~CPdf
is called.

Most of these objects keep weak_ptr to CPdf and so they are no longer able
to call weak_ptr::lock (as they are called from destructor or later
context).

This change splits CPdf destructor into 2 parts:
	- cleanup done in invalidate method - all properties and object
	  cache related stuff
	- destructor itself which takes care of the rest (like xref
	  destroying and pdf id releasing)
This will enable to call CPdf::invalidate before destroying CPdf in
PdfFileDeleter.

Please note that this change doesn't enable using weak_ptr::lock in other
cobjects destructors because we _cannot_ guarantee that there are no other
references to them outside from CPdf! So they can be destroyed anytime later
after CPdf is gone!

This change is merely cleanup which forces right destroy ordering when there
are no other references from outside.


* CPdf::invalidate added
	- all cleanup stuff moved from destructor
* CPdf::~CPdf
	- doesn't cleanup internal data structure
	- deletes only xref and releases pdf id
* PdfFileDeleter
	- calls CPdf::invalidate before deleting CPdf instance