Introduce PdfDocumentWriter
Delinearizator contains a lot of code that could be used for other purposes
as well - e.g. when we want to create what-ever new document - because every
code doing this would need to write the PDF header, some content and finally
xref with trailer.

PdfDocumentWriter is such a generalization which provides all pieces from
Delinearizator that can be reused. The basic idea is that the only think
that we need is to provide a list of objects that should be written. This is
done in abstract method fillObjectList which is, like read system function
called unless it returns 0. Therefore writeDocument method needs just to
make all sanity checks, write header, write all objects returned from
fillObjectList and finally write trailer data.

In the end we can easily migrate Delinearizator to inherit from
PdfDocumentWriter and take care only for object selection (this will be done
in later patch for simpler review).

* FileStreamData struct added
	- contains all data that are needed for file streams (currently
	  stream and file handle).
* FileStreamDataDeleter class added
	- this class should be used to destroy data from FileStreamData
	  structure. It is not used directly by PdfDocumentWriter because it
	  doesn't care about that but should be used by descendants which
	  have to do the cleanup.
	- Note that this class cannot be used from descendants' destructor
	  because it closes file handle and this can be potentially used
	  later when destroying stream in CXref::~CXref
* PdfDocumentWriter abstract class added
	- inherits from CXref
	- all the code (except for linearized data) is used from
	  Delinearizator
* PdfDocumentWriter::getStreamData
	- new method maps to initial part of Delinearizator::getInstance
	  which creates file handle and FileStreamWriter used for
	  Delinearizator constructor. We are grouping those two variables
	  into FileStreamData structure for easier usage (one return value
	  rather than 2 indirect parameters)
* PdfDocumentWriter::fillObjectList
	- should be mapped to the part of Delinearizator::delinearize(FILE*)
	  which collects objects to be written
* PdfDocumentWriter::writeDocument(const char *)
	- maps to Delinearizator::delinearize(const char*) - creates FILE
	  handle and delegates to writeDocument(FILE*)
* PdfDocumentWriter::writeDocument(FILE*)
	- maps to Delinearizator::delinearize(FILE*)
	- makes all sanity checks
	- uses IPdfWriter::writeHeader instead of hardcoded header
	  parsing&writting
	- uses fillObjectList to get all objects to be written (repeatedly
	  until no object is returned)
	- outputStream is wrapped by smart pointer to make exception
	  handling easier