Update for xslt style-sheet location
With the current update of the docbook-xsl package (to 1.74.3+dfsg-1) path
of the style-sheets is changed from
/usr/share/xml/docbook/stylesheet/nwalsh/ to
/usr/share/xml/docbook/stylesheet/docbook-xsl/. The old directory will be
linked to the old one for some time but there is already a plane to drop to
old one later.

We can keep the old path as only the correct one is chosen by configure
script which goes through all of them.