Rename Save copy to Copy revision
As LEGOManiac writes in bt#301 the "Save copy" function is rather
unintuitive because many users expect that this will make a copy which
includes also changes made to the document when done on the latest revision.

The original idea behind this functionality was to clone current revision.
Changes shouldn't be cloned because they don't form any revision at the
moment. They _will_ form one after they are saved as a new revision.

"Save current revision" expreses the functionality much better IMO.

* menu_file.xml
	- doc updated
* menu.qs
	- func_savecopy renamed to func_clonerevision
* pdfeditrc
	- fsavecopy renamed to fcopyrevision
	- Save copy... menu item renamed to Copy current revision