Revision numbers should be stable
This patch implements request from bt#310

LEGOManiac makes really a good point when he says:
---
" Intuitively, if I create a document, I would expect that newly created
document to be revision 0.
If I change it, that would be revision 0000001.
If I change it again, that would be revision 0000002, etc. etc.
Yet, when I change a document with PDFEdit, the ORIGINAL document becomes
revision 1 and the revised version becomes revision 0.

Is it supposed to be that way? Isn't that risky?
I mean, consider this:
If I edit a document for my boss and send it to him, he could potentially call
me on it later. But what if, by then, I've revised the document two more times.
In discussing the document, he would be talking about revision 0 while on my
system it would have become revision 2. That can get confusing.
I would think it would make more sense for the original document to remain
revision 0 for all time, as would revsion 1, etc. etc.
That way, if two people disuss a document and one person's has been more
recently updated than the others', at least when they both discuss revision 2,
it would be the same copy. "
---

At this moment we are comparing the current revision simply to 0. This test
is hardcoded on many places. This patch replaces these hardcoded conditions
by isLatestRevision helper function which abstracts revision numbering details.

revisions array needs to be organized in opposite ordering (to make the same
indexing logic).

Besides that we need to change all comments which are describing revision
numbering. This includes doxygen comments as well as docbook documentation.

Finally, we have to update current revision whenever we save changes as the
new revision. We didn't have to do that before, because revision==0 was kept
also after saving.


* cpdf.{cc,h}
	- doxygen doc updated
	- getActualRevision checks replaced by isLatestRevision

* xrefwriter.{cc,h}
	- isLatestRevision added
	- direct revision checks replaced by isLatestRevision
	- doxygen doc updated
	- saveChanges
		- add the current xrefPos to the end of revisions
		- revision is updated to the new value
	- collectRevisions
		- organizes revisions array in reverse order
		- set the current revision to the highest one
	- getRevisionSize
		- sync with reverse order of revisions

* testcpdf.cc
	- sync with reverse ordering of revisions
	- TC09 for revisionTC
		- check for correct changing of the latest revision