Keep changes after revision change
Currently we are dropping all changes when the revision is changed. This is
due to XRefWriter::changeRevision calling CXref::reopen which ends up
calling CXref::cleanUp method.

We cannot simply remove cleanUp from reopen because reopen is used also when
we are saving the new revision and we should drop all changes in that case
as we are basically starting a clean revision.

The easiest thing to do is to enhance reopen method signature to contain
dropChanges flag which would control cleanUp method calling.

* CXref::reopen
	- signature changed to contain dropChanges (with true default value
	  to keep changes as small as possible)
	- cleanUp is called on if dropChanges is set

* XRefWriter::changeRevision
	- calls reopen with dropChanges=false to keep all changes.

* testcpdf.cc
	- TC10 added to revisionsTC which checks the new functionality