From: mstsxfx@gmail.com To: pdfedit-devel@lists.sourceforge.net Subject: [PATCH] Enable/disable logic for Down button in MergeDialog fix
This fixes bt#315.

Currently we are using pageCount for enable/disable logic for Down button.
The problem is that we don't update pageCount when we are adding/removing
pages to the merge list. This results in faulty (premature) disabling Down
button for all pages inserted at the end.

We have to update pageCount in both addBtn_clicked and removeBtn_clicked and
we have to do it right before mergeList_currentChange method is triggered
(mergeList->setSelected method is called).

This patch also fixes some off-by-one errors in downBtn_clicked and
mergeList_currentChange. currentItem returns values from 0 while pageCount
contains total number of pages (from 1).

* mergeList_currentChange, downBtn_clicked
	- off-by-one error fixed
* addBtn_clicked, removeBtn_clicked
	- update pageCount