MergeDialog add button enable logic fix
This fixes bt#316

QT3 QListBox::insertItem has a bug which causes that the currentItem (which
is NULL for the empty list) is set when we are adding the first item without
currentChanged being notified. This means that following setCurrentItem is
NOOP because the current is already set.

This bug results in buggy behavior when we are returning page back to the
fileList which is empty. addBtn doesn't get enabled and so we cannot add the
page back. If we have a case when both original and opened files have 1 page
then we have to do some clicking exercises to force current change.

* MergeDialog::removeBtn_clicked
	- force fileList_currentChanged after setCurrentItem