CPdf::trailer cleanup
This is just cleanup change to remove CPdf::trailer field to make further
changes to support trailer editing easier.

CPdf::trailer is currently used only to initialize docCatalog and in fact it
is not required for it because XRef exports methods getRootNum and
getRootGen to get this reference.

Thus this field is used only in CPdf::getTrailer method and this one can
return the fresh (from XRef::getTrailerDict()) everytime this is called.

The advantage of this approach is that we don't have to keep XRef
trailer and CPdf::trailer in sync when we enable trailer editing.