XPDF: trailerDict access rewrite
XPdf code currently relies on the single parsed trailer which represents the
most recent revision (stored in trailerDict). All other trailers are parsed
only temporarily and only used to get Prev value for older revision.

This approach works fine also for revision changing because we are basically
doing:

CPdf::changeRevision(revision)
	XRefWriter::changeRevision(revision)
		lookup pos_of_revision
		CXref::reopen(pos_of_revision, false)
			XRef::destroyInternals
				entries cleanup
				trailerDict.free
			XRef::initInternals(pos_of_revision)
				XRef::readXRef(pos_of_revision)
					readXRefTable  - parse and initialize trailerDict
					readXRefStream -

So the XRef can see only that trailerDict which is the most recent for the
currently selected revision. So far so good.

Nevertheless, if we want to enable making changes to the trailer then this
won't work because whenever we change the revision we will reparse the
trailer when we get back and so we will lose changes already done to this
object.

The idea how to handle this without many changes to the xpdf is something
like we do for changed objects and XRef::entries (we are not changing this
XRef field directly and we are rather overriding the fetch method which finds
out whether the object is new and delegate to XRef::fetch only if it is
not).

To accomplish this we have to convert all direct uses of trailerDict to
getTrailerDict() method calls to make sure that the caller gets always the
correct instance. To ensure this I have made this field private and
converted all (but initialization) uses in XRef.cc.

One another and maybe not so obvious thing to do is to remove rootNum and
rootGen fields because they strongly depend on the current trailer which we
suppose can be editable later. All users of those fields must be changed to
use getRootNum, getRootGen methods which are already available in XRef (they
just need to get the fresh Root reference).

This shouldn't change anything in the sole xpdf code workflow because
XRef::getTrailerDict basically returns the pointer to trailerDict. In the
higher-level code, where we will want to enable changes to trailer, we just
override getTrailerDict and we are done.