CPdf: possible double fclose and mem leak in CPdf::getInstance
Fix for bt#320

File opened for the document cannot be closed unconditionally in the catch
block. The problem is that there are more possible sources of exceception.

The first one is CPdf constructor. If this one throws then there is nobody
else to close the file.
All later places have already instance created and so we cannot close the
file handle as it is closed by PdfFileDeleter.

Therefore we can (and have to) close the file handle only if the instance is
not initialized yet. The similar applies to stream which is not deallocated
if we throw in exception.