Make trailer editable
Even though CXref implements changeTrailer method this doesn't work in
general.

The main problem is that this function modifies XRef::trailerDict which is
reparsed each time when XRef::initInternals method is called what happens
everytime we change the revision so we would loose our changes.

One of the previous changes in XRef, however, changed access to
XRef::trailerDict via getTrailerDict only and so we can override this method
in CXRef and XRefWriter so that we can bypass this restriction.

The idea is simple. We will create a new trailer dictionary
CXref::currTrailer which will be allocated on demand when we make the first
change. If the currTrailer is not initialized we just pass to
XRef::getTrailerDict. This method is virtual so this will work also on xpdf
layer.

Trailer changing is organized into 3 layers (same like standard object
changing). CXref is responsible for doing the job, XRefWriter checks whether
the change is allowed (we are in the latest revision, change is safe,
document is not read-only, etc...) and CPdf makes CObject interface.

More to the change safety. There are two things to check. The first one is
whether it is safe the change the field at all (e.g. Size or Prev). Changing
them may cause total damage of the document. This check is done in
canChangeTrailerEntry util function.

The other check is for type safaty of the change. PDF specification make
rather strict restrictions about types of values for some entries (well all
defined in standard but everybody is free to add any other fields with
arbitrary types). Those specific entries are checked in typeSafeTrailerEntry
util function. Both checks are done, as already mentioned, in
XRefWriter::changeTrailer method.


* CXref
	- currTrailer added - tracks changed trailer
	- getTrailerDict overriden to return currTrailer if it is
	  initialized
	- changeTrailer - initializes currTrailer if not done yet. Still
	  relies on getTrailerDict to get proper trailer dictionary
	- cleanUp - clears currTrailer
* XRefWriter
	- canChangeTrailerEntry, typeSafeTrailerEntry implemented - also
	  exported to the API
	- getTrailerDict overriden to make latest revision check
	- changeTrailer replaces the original typeSafe check by
	  canChangeTrailerEntry and typeSafeTrailerEntry
* CPdf
	- changeTrailer added
	- getTrailer - set pdf to the CDict for cases when we need to
	  dereference indirect objects
* testcpdf
	- changeTrailerTC implemented