flattener: bad object freeing in collectReachableRefs fail path
We are currently freeing obj (input parameter) if we fail with reference
fetching. This is obvious not right thing to do. It was probably meant to
free target instead but this is already covered by shared_ptr logic.

This is not very dangerous because we are throwing an exception in that path
so we would free an object which would be freed later in the unrolling stack
(note that we are not deallocating this object).